Add the ability to whitelist by path
In some cases, a package is relied upon by multiple other packages.
In those cases, some of the base packages may be of little concern --
for example, if it's used only for testing. When that happens, we want
to whitelist that specific usage, but not other uses or new usages that
may cause more major concerns

Usage:
 - When the scan runs and a vulnerability is found, it will print out
   the path of each matching node dependency
 - The whitelist file now supports an optional array of
   "dependencyPaths"; they can be the entire path or a regex
 - When scanning and using a whitelist, if a vulnerability is found
   any paths that match a whitelisted dependency path will be ignored